First Successful Evaluation, if renamed to if*
Turns out you can’t have functions declared with the same name as built
in functions. In retrospect, this should have been obvious. Renamed if
to if* to solve that problem. Discovered an instance of M_boolean with
the wrong inputs inside of if*, corrected that. Made M_state correctly
handle the last line of a program.

Program evaluates correctly on a file of

var x;
x = 10;
var y = 3 * x + 5;
while (y % x != 3)
  y = y + 1;
if (x > y)
  return x;
else if (x * x > y)
  return x * x;
else if (x * (x + x) > y)
  return x * (x + x);
else
  return y - 1;

Success!